adding corporateGroup type: context | provider | i18n
