+++ packages/core/src/types/common/i18n/corporateGroup.ts
@@ -1,0 +1,75 @@
+import { I18nRemoveReason } from '../interfaces';
+
+export type CorporateGroupI18nProps = {
+	preview: {
+		buttons: {
+			one: string;
+			two: string;
+			three: string;
+			four: string;
+			five: string;
+		};
+		checkboxLabel: string;
+		trusteeType: string;
+	};
+	name: {
+		title: string;
+		fields: {
+			title: {
+				label: string;
+			};
+			firstName: {
+				label: string;
+				error: string;
+			};
+			lastName: {
+				label: string;
+				error: string;
+			};
+		};
+	};
+	contacts: {
+		title: string;
+		subtitle: string;
+		fields: {
+			telephone: {
+				label: string;
+				error: string;
+			};
+			email: {
+				label: string;
+				error: string;
+			};
+		};
+	};
+	professional: {
+		title: string;
+		subtitle: string;
+		fields: {
+			isProfessional: {
+				labels: {
+					isProfessionalYes: string;
+					isProfessionalNo: string;
+				};
+			};
+		};
+	};
+	remove: {
+		confirm: {
+			title: string;
+			subtitle: string;
+			breadcrumbs: {
+				link1: string;
+				link2: string;
+			};
+			dialog: {
+				message1: string;
+			};
+			buttons: {
+				remove: string;
+				cancel: string;
+			};
+		};
+		reason: I18nRemoveReason;
+	};
+};
