+++ packages/core/src/types/common/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './employer';
 export * from './trustee';
 export * from './actuary';
+export * from './corporateGroup';
