+++ packages/layout/src/components/cards/corporateGroup/i18n.ts
@@ -1,78 +1,4 @@
-import { I18nRemoveReason } from '../common/interfaces';
-
-export type CorporateGroupI18nProps = {
-	preview: {
-		buttons: {
-			one: string;
-			two: string;
-			three: string;
-			four: string;
-			five: string;
-		};
-		checkboxLabel: string;
-		trusteeType: string;
-	};
-	name: {
-		title: string;
-		fields: {
-			title: {
-				label: string;
-			};
-			firstName: {
-				label: string;
-				error: string;
-			};
-			lastName: {
-				label: string;
-				error: string;
-			};
-		};
-	};
-	contacts: {
-		title: string;
-		subtitle: string;
-		fields: {
-			telephone: {
-				label: string;
-				error: string;
-			};
-			email: {
-				label: string;
-				error: string;
-			};
-		};
-	};
-	professional: {
-		title: string;
-		subtitle: string;
-		fields: {
-			isProfessional: {
-				labels: {
-					isProfessionalYes: string;
-					isProfessionalNo: string;
-				};
-			};
-		};
-	};
-	remove: {
-		confirm: {
-			title: string;
-			subtitle: string;
-			breadcrumbs: {
-				link1: string;
-				link2: string;
-			};
-			dialog: {
-				message1: string;
-			};
-			buttons: {
-				remove: string;
-				cancel: string;
-			};
-		};
-		reason: I18nRemoveReason;
-	};
-};
+import { CorporateGroupI18nProps } from '@tpr/core';
 
 export const i18n: CorporateGroupI18nProps = {
 	preview: {
