+++ packages/layout/src/components/cards/corporateGroup/views/professional/professional.tsx
@@ -1,12 +1,15 @@
 import React, { useState } from 'react';
 import { Form, renderFields, validate, FieldProps } from '@tpr/forms';
-import { Flex } from '@tpr/core';
 import { Content } from '../../../components/content';
 import { Footer } from '../../../components/card';
 import { ArrowButton } from '../../../../buttons/buttons';
 import { useCorporateGroupContext } from '../../context';
-import { CorporateGroupI18nProps } from '../../i18n';
-import { cardType, RecursivePartial } from '../../../common/interfaces';
+import {
+	Flex,
+	cardType,
+	RecursivePartial,
+	CorporateGroupI18nProps,
+} from '@tpr/core';
 
 const getProfessionalFields = (
 	labels: RecursivePartial<CorporateGroupI18nProps['professional']['fields']>,
